requirements.txt
customtkinter
pyserial
matplotlib
meshtastic
pandas
tkintermapview
Pillow